+++ src/main/kotlin/com/zepben/ewb/cim/iec61970/base/core/ConductingEquipment.kt
@@ -99,7 +99,7 @@ abstract class ConductingEquipment(mRID: String) : Equipment(mRID) {
      * @param terminal The [Terminal] to remove.
      * @return true if [terminal] is removed from the collection.
      */
-    fun removeTerminal(terminal: Terminal): Boolean = _terminals.remove(terminal)
+    fun removeTerminal(terminal: Terminal): Boolean = _terminals.remove(terminal).also { terminal._conductingEquipment = null }
 
     /**
      * Clear all [Terminal]'s from this [ConductingEquipment].
@@ -107,7 +107,9 @@ abstract class ConductingEquipment(mRID: String) : Equipment(mRID) {
      * @return This [ConductingEquipment] for fluent use.
      */
     fun clearTerminals(): ConductingEquipment {
+        val toClearAssociation = _terminals.toList()
         _terminals.clear()
+        toClearAssociation.forEach { terminal -> terminal._conductingEquipment = null }
         return this
     }
 
@@ -116,7 +118,7 @@ abstract class ConductingEquipment(mRID: String) : Equipment(mRID) {
             return true
 
         if (terminal.conductingEquipment == null)
-            terminal.conductingEquipment = this
+            terminal._conductingEquipment = this
 
         require(terminal.conductingEquipment === this) {
             "${terminal.typeNameAndMRID()} `conductingEquipment` property references ${terminal.conductingEquipment!!.typeNameAndMRID()}, expected ${typeNameAndMRID()}."