+++ src/test/kotlin/com/zepben/ewb/cim/iec61970/base/core/ConductingEquipmentTest.kt
@@ -66,6 +66,47 @@ internal class ConductingEquipmentTest {
     }
 
     @Test
+    internal fun rejectsAddingTerminalBelongingToAnotherEquipment() {
+        val ce1 = object : ConductingEquipment(generateId()) {}
+        val ce2 = object : ConductingEquipment(generateId()) {}
+        val terminal = Terminal(generateId()).apply { ce2.addTerminal(this) }
+
+        expect { ce1.addTerminal(terminal) }
+            .toThrow<IllegalArgumentException>()
+            .withMessage(
+                "${terminal.typeNameAndMRID()} `conductingEquipment` property references ${terminal.conductingEquipment!!.typeNameAndMRID()}, expected ${ce1.typeNameAndMRID()}."
+            )
+    }
+
+    @Test
+    internal fun addingTerminalAssignsTheEquipmentToTheTerminal() {
+        val ce1 = object : ConductingEquipment(generateId()) {}
+        val terminal = Terminal(generateId()).apply { ce1.addTerminal(this) }
+
+        assertThat(terminal.conductingEquipment, equalTo(ce1))
+    }
+
+    @Test
+    internal fun removingTerminalFromTheEquipmentSetTerminalConductingEquipmentToNull() {
+        val ce1 = object : ConductingEquipment(generateId()) {}
+        val terminal = Terminal(generateId()).apply { ce1.addTerminal(this) }
+        ce1.removeTerminal(terminal)
+
+        assertThat(terminal.conductingEquipment, equalTo(null))
+    }
+
+    @Test
+    internal fun clearTerminalFromTheEquipmentSetAllTerminalConductingEquipmentToNull() {
+        val ce1 = object : ConductingEquipment(generateId()) {}
+        val terminal1 = Terminal(generateId()).apply { ce1.addTerminal(this) }
+        val terminal2 = Terminal(generateId()).apply { ce1.addTerminal(this) }
+        ce1.clearTerminals()
+
+        assertThat(terminal1.conductingEquipment, equalTo(null))
+        assertThat(terminal2.conductingEquipment, equalTo(null))
+    }
+
+    @Test
     internal fun terminals() {
         PrivateCollectionValidator.validateOrdered(
             { id -> object : ConductingEquipment(id) {} },